fix(stats): preserve dtype in no-alpha volume-budget fallback
pd.Series(0, index=loc.index) inferred int64; np.sqrt() of an integer
array always upgrades to float64, so the Gaussian (no-skew) fallback in
scale_team_volume_to_budget silently promoted the whole per-team
computation to float64. Writing that back into the float32 loc/scale/std
columns (float32 from the PyTorch-backed model output) tripped pandas
2.2's incompatible-dtype FutureWarning on every gameday during meditate's
training-data gather.

Match the fallback's dtype to loc's instead, and add a float32
regression test that fails red on the old code.

+++ src/sportstradamus/stats/base.py
@@ -297,7 +297,7 @@ def scale_team_volume_to_budget(
         sn_alpha = (
             team_df[f"proj {market} alpha"].copy()
             if f"proj {market} alpha" in team_df.columns
-            else pd.Series(0, index=loc.index)
+            else pd.Series(0, index=loc.index, dtype=loc.dtype)
         )
         N = len(team_df)
